refactor(daemon): split server.rs into daemon/* + clone audit
Tidy First extraction of `src/server.rs` (2700+ lines) into focused
modules under `src/daemon/`:

- daemon/state.rs       (143)  ConnectedClient, DragState, ClientMsg, TabAction
- daemon/router.rs      (159)  client_reader, accept_client, effective_size
- daemon/snapshot.rs    (57)   capture_workspace
- daemon/render.rs      (184)  render_frame_to_buf
- daemon/dispatch.rs    (552)  process_event, process_mouse, execute_command
- daemon/keys.rs        (655)  process_key (split out of dispatch.rs)
- daemon/event_loop.rs  (1225) run() main loop
- daemon/mod.rs         (23)   re-exports

src/server.rs is now a 16-line stub re-exporting `daemon::run`.

Cross-module call sites that previously used `super::X` to reach symbols
that lived in main.rs are updated to `crate::app::*::X` (cli/, app/ from
#23). 18 `.clone()` calls in the moved code are annotated `[perf:cold]`,
`[perf:init]`, or `[perf:hot]` with rationale; 3 `TODO(perf)` flags mark
the Layout::clone() and per-mouse-event cache clone for follow-up Arc
conversion (separate issue).

Behavior-preserving: 105 unit + 3 integration + 8 property + 4 daemon
lifecycle tests pass. Bench within ±2% of baseline (no regression).

MAINTENANCE.md gains a "Performance profiling" section with flamegraph
instructions.

Closes #24
